Crate steamworks

Futures-enabled bindings to a tiny portion of the Steamworks API.

You will probably want to keep the official Steamworks Documentation open while reading these API docs, as it contains a lot of information which is not documented here.

The Client::init function will initialize the Steamworks API, and provide the Client object, which provides the Steamworks API functionality. Note that for initialization to succeed, the Steam client needs to be running and you'll probably need to create a steam_appid.txt file; see this section for the full details.

Example

use steamworks::Client;

let client = Client::init()?;

// Gets the App ID of our application
let app_id = client.app_id();

Modules

callbacks

remote_storage

ugc

user_stats

Structs

AppId

Client
The core type of this crate, representing an initialized Steamworks API.

SteamId

Enums

InitError

SteamResult