Struct steamworks::user_stats::LeaderboardHandle

pub struct LeaderboardHandle { /* fields omitted */ }

A handle to a Steam leaderboard

The functions on this handle wrap the DownloadLeaderboardEntries() and GetDownloadedLeaderboardEntry() Steamworks API functions.

Implementations

impl LeaderboardHandle

pub fn download_global(
&self,
range_start: u32,
range_end: u32,
max_details: u8
) -> impl Future<Output = Vec<LeaderboardEntry>> + Send + '_

Fetches a sequential range of leaderboard entries by global rank.

range_start and range_end are both inclusive. max_details should be 64 or less; higher values will be clamped.

Panics

Panics if range_start < 1 or range_end < range_start.

pub fn download_global_around_user(
&self,
range_start: i32,
range_end: i32,
max_details: u8
) -> impl Future<Output = Vec<LeaderboardEntry>> + Send + '_

Fetches a sequential range of leaderboard entries by position relative to the current user's rank.

range_start and range_end are both inclusive. max_details should be 64 or less; higher values will be clamped.

Panics

Panics if range_end < range_start.

pub fn download_friends(
&self,
max_details: u8
) -> impl Future<Output = Vec<LeaderboardEntry>> + Send + '_

Fetches all leaderboard entries for friends of the current user.

max_details should be 64 or less; higher values will be clamped.

pub fn upload_leaderboard_score<'a>(
&'a self,
score: i32,
details: Option<&'a [i32]>,
force_update: bool
) -> impl Future<Output = Result<LeaderboardScoreUploaded, UploadLeaderboardScoreError>> + Send + 'a

Uploads a score to the leaderboard.

details is optional game-specific information to upload along with the score. If force_update is true, the user's score is updated to the new value, even if the new score is not better than the already existing score (where "better" is defined by the leaderboard sort method).

Panics

Panics if details, if provided, has a length greater than 64.